fix(poll): Callback handle argument type.

+++ core/epoll.c
@@ -186,7 +186,7 @@ nil_t poll_call_usr_on_open(poll_p poll, i64_t id) {
 
     // Call the callback if it's a lambda
     if (clbfn != NULL && (*clbfn)->type == TYPE_LAMBDA) {
-        stack_push(i32(id));
+        stack_push(i64(id));
         v = call(*clbfn, 1);
         if (IS_ERROR(v)) {
             f = obj_fmt(v, B8_FALSE);
@@ -210,7 +210,7 @@ nil_t poll_call_usr_on_close(poll_p poll, i64_t id) {
 
     // Call the callback if it's a lambda
     if (clbfn != NULL && (*clbfn)->type == TYPE_LAMBDA) {
-        stack_push(i32(id));
+        stack_push(i64(id));
         v = call(*clbfn, 1);
         if (IS_ERROR(v)) {
             f = obj_fmt(v, B8_FALSE);
